More bug fixes
1) Brace matching tagger threw exception when content of file had
outlines but then they were removed (i.e. clearing a file).
2) Incomplete strings (while typing) were causing function parsing to
fail. Still need to figure out multi-line strings, but this is not high
priority right now.
3) Adjustments to the parser to make define, type, and constant def
analysis determine if the block is incomplete (needed for completion
analysis)
4) Added a few new keywords
5) Fixed the drop-down function list client so that if more than one
window of the same document is opened, the function list will appear in
subsequent windows.