Prevent "key already added" exceptions
There are a few places in the parser that use non-concurrent
dictionaries, so for those, we're now checking to see if the key exists
before attempting to add. Also did a bunch of this in the xml settings
loader, just in case things somehow get screwed up in the xml file.