Added API for env-specific commands
The Genero 4GL EditFilter class handles all commands that come in
through the text editor. However, there are some cases when those
commands need to be further filtered and perhaps executed. Added a new
API to allow another package to register for the commands coming through
a .4gl file and optionally do something with them.